cloudrun/requirements.txt
flask==2.3.3
gunicorn==21.2.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.0
requests==2.31.0
python-dotenv==1.0.0
flask-session==0.5.0
cachelib==0.9.0

